Minor edge case updates

+++ sign_int_fmt.c
@@ -9,7 +9,7 @@
  */
 char *sign_int_fmt(va_list args, FMT *fmt)
 {
-	unsigned long n = va_arg(args, long);
+	unsigned long n = va_arg(args, unsigned int);
 	int i = 1;
 	char *num = malloc(i + 1);
